feat(docs): enhance changelog functionality and layout
- Wrapped children in NuqsAdapter within the Layout component for improved context handling.
- Refactored ChangelogPage to simplify changelog file reading.
- Introduced ChangelogEntryItem component for better modularity in displaying changelog entries.
- Updated ChangelogViewer to utilize the new ChangelogEntryItem component and refined package filtering logic.
- Improved parsing logic in parse-changelog to enhance changelog entry structure documentation.
